Added a size type specific to ScriptObjectKeys size and count

+++ src/libscriptobject/script_object_keys.cpp
@@ -24,7 +24,7 @@ void rs::scriptobject::ScriptObjectKeys::ScriptObjectKeysDeleter(ScriptObjectKey
     delete[] reinterpret_cast<char*>(ptr); 
 }
 
-rs::scriptobject::ScriptObjectKeys::ScriptObjectKeys(unsigned short s, unsigned short c, const ScriptObjectHash h) : size(s), count(c) {
+rs::scriptobject::ScriptObjectKeys::ScriptObjectKeys(size_t s, size_t c, const ScriptObjectHash h) : size(s), count(c) {
     std::copy_n(h, sizeof(hash), hash);
 }
